trace-cmd: Fix writing of uncompressed size
pointer in struct tracecmd_compression is 'unsigned long', which is 8 byte
in size on most platforms, but the tep_read_number() call in the next line
treats it as a 4 byte value. As there's no need for unsigned long change
the type to unsigned int.

Link: https://lore.kernel.org/all/20220711074418.858843-1-svens@linux.ibm.com/
Link: https://lore.kernel.org/all/20220711094340.2829115-1-svens@linux.ibm.com/
Link: https://lore.kernel.org/linux-trace-devel/20220711142104.090b9030@gandalf.local.home

Fixes: 3f8447b1 ("trace-cmd library: Add support for compression algorithms")
Acked-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>

+++ lib/trace-cmd/trace-compress.c
@@ -32,7 +32,7 @@ struct tracecmd_compression {
 	int				fd;
 	unsigned int			capacity;
 	unsigned int			capacity_read;
-	unsigned long			pointer;
+	unsigned int			pointer;
 	char				*buffer;
 	struct compress_proto		*proto;
 	struct tep_handle		*tep;
